requirements.txt
Flask
Flask_Marshmallow
Flask_Sqlalchemy
marshmallow-sqlalchemy

